+++ src/inspect_ai/model/_openai.py
@@ -61,8 +61,7 @@ from inspect_ai._util.http import (
     is_retryable_http_status,
     parse_retry_after_from_exception,
 )
-from inspect_ai._util.images import file_as_data_uri
-from inspect_ai._util.url import is_http_url
+from inspect_ai._util.images import inline_media_data_uri
 from inspect_ai.model._call_tools import parse_tool_call
 from inspect_ai.model._generate_config import GenerateConfig
 from inspect_ai.model._internal import (
@@ -192,13 +191,8 @@ async def openai_chat_completion_part(
     if content.type == "text":
         return ChatCompletionContentPartTextParam(type="text", text=content.text)
     elif content.type == "image":
-        # API takes URL or base64 encoded file. If it's a remote file or
-        # data URL leave it alone, otherwise encode it
-        image_url = content.image
+        image_url = inline_media_data_uri(content.image, "image")
         detail = content.detail
-
-        if not is_http_url(image_url):
-            image_url = await file_as_data_uri(image_url)
 
         return ChatCompletionContentPartImageParam(
             type="image_url",
@@ -207,14 +201,14 @@ async def openai_chat_completion_part(
             ),
         )
     elif content.type == "audio":
-        audio_data_uri = await file_as_data_uri(content.audio)
+        audio_data_uri = inline_media_data_uri(content.audio, "audio")
         audio_data = audio_data_uri.split("base64,")[1]
 
         return ChatCompletionContentPartInputAudioParam(
             type="input_audio", input_audio=dict(data=audio_data, format=content.format)
         )
     elif content.type == "document":
-        document_data_uri = await file_as_data_uri(content.document)
+        document_data_uri = inline_media_data_uri(content.document, "document")
 
         return File(
             type="file",